Back to unsigned int count in wxTreeCtrl.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@38351 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ include/wx/generic/treectlg.h
@@ -69,7 +69,7 @@ public:
     // implement base class pure virtuals
     // ----------------------------------
 
-    virtual size_t GetCount() const;
+    virtual unsigned int GetCount() const;
 
     virtual unsigned int GetIndent() const { return m_indent; }
     virtual void SetIndent(unsigned int indent);
