+++ src/mac/carbon/treectrl.cpp
@@ -57,7 +57,7 @@ wxTreeCtrl::~wxTreeCtrl()
 }
 
 // Attributes
-size_t wxTreeCtrl::GetCount() const
+unsigned int wxTreeCtrl::GetCount() const
 {
     // TODO
     return 0;
