+++ src/mac/classic/treectrl.cpp
@@ -55,7 +55,7 @@ wxTreeCtrl::~wxTreeCtrl()
 }
 
 // Attributes
-size_t wxTreeCtrl::GetCount() const
+unsigned int wxTreeCtrl::GetCount() const
 {
     // TODO
     return 0;
